create football api module
